...and are now scoped by account ID.

+++ src/gateway/sql/static/ansi/keys/delete.sql
@@ -1,2 +1,3 @@
 DELETE FROM keys
 WHERE keys.id = ?
+  AND keys.account_id = ?
